requirements.txt
ipython>=1.1.0
mysql-python
pytz

listparser

django>=1.6.1
South>=0.8.4
django-extensions
django-admin-tools
django-bootstrap3
django-bootstrap-pagination
django-grappelli
django-taggit
django-taggit-live

django-setuptest
argparse
